RIP Mercurial
Hg introduced me to the DVCS world. It provided the initial version
control for this repo. When it hit the scene, I was enamored with Python
and the difference from subversion was night and day. Years later, I
picked up git for work, learned to love that everything accepts a
tree-ish, and never looked back. This repo migrated some time in 2015
and, since then, I've rarely touched hg. I dropped the vim integrations
some time ago. Time to jettison the remnants.